requirements.txt
pyserial
picamera